requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiohttp-jinja2==1.6
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
frozenlist==1.6.0
idna==3.10
Jinja2==3.1.6
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.4.4
paramiko==3.5.1
platformdirs==4.3.8
propcache==0.3.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyNaCl==1.5.0
PyYAML==6.0.2
rich==14.0.0
setuptools==80.9.0
textual==3.3.0
textual-dev==1.7.0
textual-serve==1.1.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uc-micro-py==1.0.3
# (vwrconf==0.1)
-e /home/pilotero/Documentos/projects/vwrconf
yarl==1.20.0
